feat!: default to async-channel entirely
