+++ Cargo.toml
@@ -26,23 +26,18 @@ log = { version = "0.4", features = ["kv"] }
 futures = { version = "0.3" }
 dyn-clone = "1.0"
 async-lock = "3"
+async-channel = "2.5.0"
 
 tokio = { version = "1", features = ["full"], optional = true }
 
 async-io = { version = "2.6.0", optional = true }
 async-global-executor = { version = "3", optional = true }
 
-async-channel = { version = "2.5.0", optional = true }
-
 [features]
 default = [
     "tokio_runtime",
     # "async_runtime",
-    # "async_channel"
 ]
-async_channel = ["dep:async-channel"]
-futures_channel = []
-
 tokio_runtime = ["runtime", "dep:tokio", "async-global-executor/tokio"]
 async_runtime = ["runtime", "dep:async-io", "dep:async-global-executor"]
