+++ justfile
@@ -13,19 +13,12 @@ clippy:
     cargo --quiet clippy --workspace --quiet --lib --tests --no-default-features --features tokio_runtime
     cargo --quiet clippy --workspace --quiet --lib --tests --no-default-features --features async_runtime
     cargo --quiet clippy --workspace --quiet --lib --tests --no-default-features --features async_runtime,tokio
-    cargo --quiet clippy --workspace --quiet --lib --tests --no-default-features --features async_channel,tokio_runtime
-    cargo --quiet clippy --workspace --quiet --lib --tests --no-default-features --features async_channel,async_runtime
-    cargo --quiet clippy --workspace --quiet --lib --tests --no-default-features --features async_channel,async_runtime,tokio
 
 test $RUST_LOG="trace" $STRESS_COUNT="3":
     cargo nextest run --profile ci --workspace --all-targets --stress-count $STRESS_COUNT
     cargo nextest run --profile ci --workspace --lib --no-default-features --features tokio_runtime --stress-count $STRESS_COUNT
     cargo nextest run --profile ci --workspace --lib --no-default-features --features async_runtime --stress-count $STRESS_COUNT
     cargo nextest run --profile ci --workspace --lib --no-default-features --features async_runtime,tokio --stress-count $STRESS_COUNT
-
-    cargo nextest run --profile ci --workspace --lib --no-default-features --features async_channel,tokio_runtime --stress-count $STRESS_COUNT
-    cargo nextest run --profile ci --workspace --lib --no-default-features --features async_channel,async_runtime --stress-count $STRESS_COUNT
-    cargo nextest run --profile ci --workspace --lib --no-default-features --features async_channel,async_runtime,tokio --stress-count $STRESS_COUNT
 
 doc_test:
     cargo test --workspace --doc --no-default-features --features tokio_runtime
@@ -39,11 +32,8 @@ install-deps:
 
 semver-checks:
     cargo semver-checks --only-explicit-features --features tokio_runtime
-    cargo semver-checks --only-explicit-features --features tokio_runtime,async_channel
     cargo semver-checks --only-explicit-features --features async_runtime
     cargo semver-checks --only-explicit-features --features async_runtime,tokio
-    cargo semver-checks --only-explicit-features --features async_runtime,tokio,async_channel
-    cargo semver-checks --only-explicit-features --features async_runtime,async_channel
 
 build-examples:
     cargo build --manifest-path hannibal-examples/Cargo.toml --features tokio_runtime
